* 오브젝트 lombok으로 모두 컨버팅. * user, summoner, target간의 상호 관계 재 조정. - 이 것이 3종이 존재해서 헷갈림. -> user 제거. * champion data API -> json으로 규격 바뀐 것 적용. - https://developer.riotgames.com/static-data.html -> https://ddragon.leagueoflegends.com/api/versions.json -> http://ddragon.leagueoflegends.com/cdn/8.18.2/data/ko_KR/champion.json

+++ src/main/java/com/lol/tracer/repository/ChampionRepository.java
@@ -3,6 +3,6 @@ package com.lol.tracer.repository;
 import com.lol.tracer.model.Champion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChampionRepository extends JpaRepository<Champion,Integer>{
+public interface ChampionRepository extends JpaRepository<Champion, Long>{
 
 }
